+++ src/Send/SendScreen/SendAllWarning.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import DangerousActionModal from '../../../legacy/components/Common/DangerousActionModal'
 import {Text} from '../../../legacy/components/UiKit'
 import {getAssetDenominationOrId, truncateWithEllipsis} from '../../../legacy/utils/format'
+import {DangerousActionModal} from '../../components'
 import {useTokenInfo} from '../../hooks'
 import {useSelectedWallet} from '../../SelectedWallet'
 import {useStrings} from './strings'
