+++ src/components/DangerousActionModal/index.ts
@@ -1,0 +1,1 @@
+export * from './DangerousActionModal'
